# docs/openapi/core.yaml
openapi: 3.0.3
info:
  title: MANI-Core — Identidad y Registro de Aliado
  description: >
    Contrato OpenAPI entre MANI-APIGateway y MANI-Node (Core) para el dominio
    de identidad: autenticación y registro de aliado persona natural.


    ## Propagación del claim de tenant

    MANI es multi-tenant: cada petición debe resolverse en el contexto de un
    tenant (organización) específico, y el Core nunca puede confiar en un
    valor de tenant que un cliente final pueda falsificar.

    1. **Pre-autenticación** (`POST /auth/login`, `POST /auth/register/ally`):
       el llamador todavía no tiene un JWT. El tenant se identifica con el
       header `X-Tenant-Id`, que **únicamente MANI-APIGateway** debe fijar
       (resuelto a partir del dominio/app-key de origen de la petición). El
       Core responde `400 TENANT_NOT_FOUND` si el tenant no existe o está
       inactivo.
    2. **Emisión del claim**: si el login o el registro son exitosos, el
       Core emite un `accessToken` (JWT) cuyo payload incluye el claim
       `tenant_id` con el id del tenant resuelto en el paso anterior (ver
       `TokenClaims`). Ese mismo id se refleja en `ProfileSummary.tenantId`
       en la respuesta.
    3. **Post-autenticación** (`POST /auth/refresh`, `POST /auth/logout`, y
       en general cualquier endpoint protegido con `bearerAuth`): el tenant
       se deriva **exclusivamente** del claim `tenant_id` dentro del JWT ya
...

...
  version: 0.2.0
servers:
...
  - url: http://localhost:3000/api/v1
...
paths:
...
  /auth/register/ally:
    post:
      tags: [Ally Registration]
      operationId: registerAllyNaturalPerson
      summary: Registra un nuevo Aliado persona natural
      description: >
        Crea el usuario y el perfil de Aliado (rol ALLY) en estado PENDING,
        a la espera de verificación KYC, dentro del tenant resuelto por
        `X-Tenant-Id`. Devuelve tokens de sesión (con el claim `tenant_id`
        ya embebido) para que el Aliado pueda continuar el flujo de
        onboarding inmediatamente.
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAllyNaturalPersonRequest'
      responses:
        '201':
          description: Aliado registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAllyResponse'
        '400':
          description: >
            Body inválido (`VALIDATION_ERROR`) o tenant inexistente/inactivo
            (`TENANT_NOT_FOUND`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            Ya existe un usuario del tenant con ese email (`EMAIL_ALREADY_REGISTERED`)
            o ese número de documento (`DOCUMENT_ALREADY_REGISTERED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        El Core verifica la firma y extrae userId (`sub`), `role` y
        `tenant_id` del payload. Ver `TokenClaims`.
  parameters:
    XTenantId:
      name: X-Tenant-Id
      in: header
      required: true
      description: >
        Id del tenant resuelto por MANI-APIGateway a partir del dominio/app-key
        de origen. Solo se usa en endpoints pre-autenticación; el Gateway es
        el único llamador autorizado a fijarlo — un cliente final nunca debe
        poder establecerlo directamente.
      schema:
        type: string
        example: trama-demo
    XCorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: >
        Id de trazabilidad distribuida propagado desde el Gateway. Si no se
        envía, el Core genera uno (`node-<timestamp>`) y lo devuelve en el
        header de respuesta y en el body de error.
      schema:
        type: string
  responses:
    InternalError:
      description: Error interno del servicio (`INTERNAL_ERROR`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorCode:
      type: string
      description: Catálogo cerrado de códigos de error del dominio de identidad.
      enum:
        - VALIDATION_ERROR
        - INVALID_CREDENTIALS
        - UNAUTHORIZED
        - TOKEN_INVALID
        - TOKEN_EXPIRED
        - TENANT_NOT_FOUND
        - EMAIL_ALREADY_REGISTERED
        - DOCUMENT_ALREADY_REGISTERED
        - INTERNAL_ERROR
      x-code-meanings:
        VALIDATION_ERROR: 'HTTP 400 — el body no cumple el esquema (campo faltante/formato inválido).'
        INVALID_CREDENTIALS: 'HTTP 401 — email y/o password no coinciden con ningún usuario del tenant.'
        UNAUTHORIZED: 'HTTP 401 — falta el header Authorization en un endpoint protegido.'
        TOKEN_INVALID: 'HTTP 401 — el JWT no tiene un formato o firma válidos.'
        TOKEN_EXPIRED: 'HTTP 401 — el JWT es válido pero ya expiró.'
        TENANT_NOT_FOUND: 'HTTP 400 — X-Tenant-Id no corresponde a un tenant existente/activo.'
        EMAIL_ALREADY_REGISTERED: 'HTTP 409 — ya existe un usuario con ese email en el tenant.'
        DOCUMENT_ALREADY_REGISTERED: 'HTTP 409 — ya existe un Aliado con ese número de documento en el tenant.'
        INTERNAL_ERROR: 'HTTP 500 — fallo no controlado del servicio.'
    ErrorResponse:
      type: object
      description: Formato estándar de error de todo el Core (ver `src/middlewares/errorHandler.middleware.js`).
      properties:
        error:
          type: string
          description: Mensaje legible, seguro para mostrar al usuario final.
          example: Credenciales inválidas
        code:
          $ref: '#/components/schemas/ErrorCode'
        correlationId:
          type: string
          example: node-1791318617328
      required: [error, correlationId]
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT de corta duración. Payload descrito en `TokenClaims`.
        refreshToken:
          type: string
          description: Token opaco de larga duración, usado solo contra `/auth/refresh` y `/auth/logout`.
        expiresIn:
          type: integer
          description: Segundos hasta la expiración del accessToken
          example: 3600
      required: [accessToken, refreshToken, expiresIn]
...
    ProfileSummary:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
          description: Refleja el claim `tenant_id` del JWT emitido para este perfil.
          example: trama-demo
        role:
          type: string
          enum: [CLIENT, ALLY, ADMIN]
        fullName:
          type: string
        status:
          type: string
          enum: [VERIFIED, PENDING, REJECTED]
      required: [id, tenantId, role, fullName, status]
...
    RegisterAllyNaturalPersonRequest:
      type: object
      description: Datos mínimos de registro para un Aliado persona natural. El tenant se toma del header `X-Tenant-Id`, no del body.
      properties:
        fullName:
          type: string
          example: María Fernanda Rojas
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        phone:
          type: string
          example: '+573001234567'
        documentType:
          type: string
          enum: [CC, CE, PASSPORT]
          description: Tipo de documento de identidad (persona natural)
        documentNumber:
          type: string
          example: '1020304050'
      required:
        - fullName
        - email
        - password
        - phone
        - documentType
        - documentNumber
    RegisterAllyResponse:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/ProfileSummary'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
      required: [profile, tokens]